refactor: move HostedTeleopRecorder out of the blueprint file
Blueprint files should define blueprints, not modules/classes (all_blueprints
even registered the recorder as a module from blueprints.py). Move
HostedTeleopRecorder/Config and HOSTED_RECORDINGS_DIR into a new
quest_hosted/recorder.py module and point the all_modules entry at it.
