+++ dimos/teleop/quest_hosted/recorder.py
@@ -1,0 +1,43 @@
+# Copyright 2025-2026 Dimensional Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Hosted-teleop recorder: the generic recorder defaulting to the hosted dir."""
+
+from pathlib import Path
+
+from dimos.constants import DIMOS_PROJECT_ROOT
+from dimos.teleop.utils.recorder import TeleopRecorder, TeleopRecorderConfig
+
+HOSTED_RECORDINGS_DIR = DIMOS_PROJECT_ROOT / "data/hosted_teleop/recordings"
+
+
+class HostedTeleopRecorderConfig(TeleopRecorderConfig):
+    # Same generic recorder, just defaulting recordings into the hosted dir.
+    db_path: str | Path = HOSTED_RECORDINGS_DIR / "recording_hosted.db"
+
+
+class HostedTeleopRecorder(TeleopRecorder):
+    """Generic ``TeleopRecorder`` defaulting to the hosted recordings dir.
+
+    Ports + per-run timestamping are inherited; this only changes the default
+    output path. Compose at the CLI::
+
+        dimos run teleop-hosted-xarm7 hosted-teleop-recorder
+        dimos run teleop-hosted-go2   hosted-teleop-recorder
+    """
+
+    config: HostedTeleopRecorderConfig
+
+
+__all__ = ["HOSTED_RECORDINGS_DIR", "HostedTeleopRecorder", "HostedTeleopRecorderConfig"]
